app/requirements.txt
fastapi==0.115.7
pydantic==2.10.6
pydantic-settings==2.7.1
uvicorn==0.34.0
alembic==1.14.1
asyncpg==0.30.0
SQLAlchemy==2.0.37
aiobotocore==2.19.0
redis==5.2.1
fastapi-cache2==0.2.2
python-multipart==0.0.20
